requirements.txt
Django==3.2.15
Pillow==10.0.0

django-allauth==0.55.2
django-debug-toolbar==4.2.0
django-environ==0.11.2
django-extensions==3.2.3
djangorestframework==3.14.0

flake8==6.1.0
isort==5.12.0

stripe==6.4.0

redis==4.6.0
celery==5.3.4
django-redis==5.3.0

psycopg2==2.9.7
django-phonenumber-field==7.1.0